+++ example/ios/RunnerTests/RunnerTests.swift
@@ -12,7 +12,7 @@ import XCTest
 class RunnerTests: XCTestCase {
 
   func testGetPlatformVersion() {
-    let plugin = FlutterQrcodePlugin()
+    let plugin = LumiQrScannerPlugin()
 
     let call = FlutterMethodCall(methodName: "getPlatformVersion", arguments: [])