+++ ios/Classes/FlutterQrcodePlugin.swift
@@ -3,10 +3,10 @@ import UIKit
 import AVFoundation
 import Vision
 
-public class FlutterQrcodePlugin: NSObject, FlutterPlugin {
+public class LumiQrScannerPlugin: NSObject, FlutterPlugin {
   public static func register(with registrar: FlutterPluginRegistrar) {
     let channel = FlutterMethodChannel(name: "lumi_qr_scanner", binaryMessenger: registrar.messenger())
-    let instance = FlutterQrcodePlugin()
+    let instance = LumiQrScannerPlugin()
     registrar.addMethodCallDelegate(instance, channel: channel)
 
     // Register platform view